used standardized integer values.

+++ server_op/socket/bind/bindsocket.cpp
@@ -1,6 +1,6 @@
 #include "./bindsocket.h"
 
-http::bindsocket::bindsocket(int domain, int service, int protocol, int port, u_long bridge)
+http::bindsocket::bindsocket(int domain, int service, int protocol, uint16_t port, uint32_t bridge)
 : dnnsocket(domain, service, protocol, port, bridge) {
   set_connection(connect_to_network(get_socket(), get_address()));
   test_connection(get_connection());
